added prompt to help performance for mass-conversions
The user is now prompted if all PD2 files are from a specific season, which will force the text files from that season to be used instead of trying multiple files and auto-detecting the season. This will reduce the amount of code that is executed, but the program won't check other season's text files for compatibility if character/stash files cannot be converted with the specified season's text files.