added s6 patch files
Added s6 patch files and updated the converter with placeholder s7/s8 files. It should convert files correctly when those placeholder files are swapped for the true versions. Note that it'll always attempt to convert to the most recent season, so both of the s7/s8 placeholder files should be replaced with the true s7 files when they become available.